Removed unecessary methods Generic bound view

+++ src/main/scala/de/thatsich/minecraft/intellie/applied/aerodynamics/proxy/module/suite/horseshoes/item/config/HorseShoesArmorConfig.scala
@@ -13,9 +13,9 @@ trait HorseShoesArmorConfig
 
 	def maximalAbsorptionRatio: Double
 
-	def minimalArmorBase: Double
+	def minimalArmorBase: Int
 
-	def maximalArmorBase: Double
+	def maximalArmorBase: Int
 
 	def minimalEnergyPerDamage: Double
